linux-1.0

+++ fs/nfs/proc.c
@@ -1,7 +1,7 @@
 /*
  *  linux/fs/nfs/proc.c
  *
- *  Copyright (C) 1992  Rick Sladkey
+ *  Copyright (C) 1992, 1993, 1994  Rick Sladkey
  *
  *  OS-independent nfs remote procedure call functions
  */
@@ -27,15 +27,37 @@
 #include <linux/in.h>
 
 #ifdef NFS_PROC_DEBUG
-static int proc_debug = 0;
-#define PRINTK if (proc_debug) printk
-#else
-#define PRINTK if (0) printk
-#endif
 
-static int *nfs_rpc_header(int *p, int procedure);
+static int proc_debug = 0;
+#define PRINTK(format, args...) \
+	do {						\
+		if (proc_debug)				\
+			printk(format , ## args);	\
+	} while (0)
+
+#else /* !NFS_PROC_DEBUG */
+
+#define PRINTK(format, args...) do ; while (0)
+
+#endif /* !NFS_PROC_DEBUG */
+
+static int *nfs_rpc_header(int *p, int procedure, int ruid);
 static int *nfs_rpc_verify(int *p);
 static int nfs_stat_to_errno(int stat);
+
+/*
+ * Our memory allocation and release functions.
+ */
+
+static inline int *nfs_rpc_alloc(void)
+{
+	return (int *) __get_free_page(GFP_KERNEL);
+}
+
+static inline void nfs_rpc_free(int *p)
+{
+	free_page((long) p);
+}
 
 /*
  * Here are a bunch of xdr encode/decode functions that convert
@@ -108,7 +130,7 @@ static inline int *xdr_decode_data(int *p, char *data, int *lenp, int maxlen)
 
 static int *xdr_decode_fattr(int *p, struct nfs_fattr *fattr)
 {
-	fattr->type = ntohl(*p++);
+	fattr->type = (enum nfs_ftype) ntohl(*p++);
 	fattr->mode = ntohl(*p++);
 	fattr->nlink = ntohl(*p++);
 	fattr->uid = ntohl(*p++);
@@ -161,18 +183,25 @@ static int *xdr_decode_fsinfo(int *p, struct nfs_fsinfo *res)
 	return p;
 }
 
+/*
+ * One function for each procedure in the NFS protocol.
+ */
+
 int nfs_proc_getattr(struct nfs_server *server, struct nfs_fh *fhandle,
 		     struct nfs_fattr *fattr)
 {
 	int *p, *p0;
 	int status;
+	int ruid = 0;
 
 	PRINTK("NFS call  getattr\n");
-	p = p0 = (int *) get_free_page(GFP_KERNEL);
-	p = nfs_rpc_header(p, NFSPROC_GETATTR);
+	if (!(p0 = nfs_rpc_alloc()))
+		return -EIO;
+retry:
+	p = nfs_rpc_header(p0, NFSPROC_GETATTR, ruid);
 	p = xdr_encode_fhandle(p, fhandle);
 	if ((status = nfs_rpc_call(server, p0, p)) < 0) {
-		free_page((long) p0);
+		nfs_rpc_free(p0);
 		return status;
 	}
 	if (!(p = nfs_rpc_verify(p0)))
@@ -181,9 +210,14 @@ int nfs_proc_getattr(struct nfs_server *server, struct nfs_fh *fhandle,
 		p = xdr_decode_fattr(p, fattr);
 		PRINTK("NFS reply getattr\n");
 	}
-	else
+	else {
+		if (!ruid && current->euid == 0 && current->uid != 0) {
+			ruid = 1;
+			goto retry;
+		}
 		PRINTK("NFS reply getattr failed = %d\n", status);
-	free_page((long) p0);
+	}
+	nfs_rpc_free(p0);
 	return -nfs_stat_to_errno(status);
 }
 
@@ -192,14 +226,17 @@ int nfs_proc_setattr(struct nfs_server *server, struct nfs_fh *fhandle,
 {
 	int *p, *p0;
 	int status;
+	int ruid = 0;
 
 	PRINTK("NFS call  setattr\n");
-	p = p0 = (int *) get_free_page(GFP_KERNEL);
-	p = nfs_rpc_header(p, NFSPROC_SETATTR);
+	if (!(p0 = nfs_rpc_alloc()))
+		return -EIO;
+retry:
+	p = nfs_rpc_header(p0, NFSPROC_SETATTR, ruid);
 	p = xdr_encode_fhandle(p, fhandle);
 	p = xdr_encode_sattr(p, sattr);
 	if ((status = nfs_rpc_call(server, p0, p)) < 0) {
-		free_page((long) p0);
+		nfs_rpc_free(p0);
 		return status;
 	}
 	if (!(p = nfs_rpc_verify(p0)))
@@ -208,9 +245,14 @@ int nfs_proc_setattr(struct nfs_server *server, struct nfs_fh *fhandle,
 		p = xdr_decode_fattr(p, fattr);
 		PRINTK("NFS reply setattr\n");
 	}
-	else
+	else {
+		if (!ruid && current->euid == 0 && current->uid != 0) {
+			ruid = 1;
+			goto retry;
+		}
 		PRINTK("NFS reply setattr failed = %d\n", status);
-	free_page((long) p0);
+	}
+	nfs_rpc_free(p0);
 	return -nfs_stat_to_errno(status);
 }
 
@@ -219,18 +261,21 @@ int nfs_proc_lookup(struct nfs_server *server, struct nfs_fh *dir, const char *n
 {
 	int *p, *p0;
 	int status;
+	int ruid = 0;
 
 	PRINTK("NFS call  lookup %s\n", name);
 #ifdef NFS_PROC_DEBUG
 	if (!strcmp(name, "xyzzy"))
 		proc_debug = 1 - proc_debug;
 #endif
-	p = p0 = (int *) get_free_page(GFP_KERNEL);
-	p = nfs_rpc_header(p, NFSPROC_LOOKUP);
+	if (!(p0 = nfs_rpc_alloc()))
+		return -EIO;
+retry:
+	p = nfs_rpc_header(p0, NFSPROC_LOOKUP, ruid);
 	p = xdr_encode_fhandle(p, dir);
 	p = xdr_encode_string(p, name);
 	if ((status = nfs_rpc_call(server, p0, p)) < 0) {
-		free_page((long) p0);
+		nfs_rpc_free(p0);
 		return status;
 	}
 	if (!(p = nfs_rpc_verify(p0)))
@@ -240,9 +285,14 @@ int nfs_proc_lookup(struct nfs_server *server, struct nfs_fh *dir, const char *n
 		p = xdr_decode_fattr(p, fattr);
 		PRINTK("NFS reply lookup\n");
 	}
-	else
+	else {
+		if (!ruid && current->euid == 0 && current->uid != 0) {
+			ruid = 1;
+			goto retry;
+		}
 		PRINTK("NFS reply lookup failed = %d\n", status);
-	free_page((long) p0);
+	}
+	nfs_rpc_free(p0);
 	return -nfs_stat_to_errno(status);
 }
 
@@ -251,13 +301,16 @@ int nfs_proc_readlink(struct nfs_server *server, struct nfs_fh *fhandle,
 {
 	int *p, *p0;
 	int status;
+	int ruid = 0;
 
 	PRINTK("NFS call  readlink\n");
-	p = p0 = (int *) get_free_page(GFP_KERNEL);
-	p = nfs_rpc_header(p, NFSPROC_READLINK);
+	if (!(p0 = nfs_rpc_alloc()))
+		return -EIO;
+retry:
+	p = nfs_rpc_header(p0, NFSPROC_READLINK, ruid);
 	p = xdr_encode_fhandle(p, fhandle);
 	if ((status = nfs_rpc_call(server, p0, p)) < 0) {
-		free_page((long) p0);
+		nfs_rpc_free(p0);
 		return status;
 	}
 	if (!(p = nfs_rpc_verify(p0)))
@@ -270,9 +323,14 @@ int nfs_proc_readlink(struct nfs_server *server, struct nfs_fh *fhandle,
 		else
 			PRINTK("NFS reply readlink %s\n", res);
 	}
-	else
+	else {
+		if (!ruid && current->euid == 0 && current->uid != 0) {
+			ruid = 1;
+			goto retry;
+		}
 		PRINTK("NFS reply readlink failed = %d\n", status);
-	free_page((long) p0);
+	}
+	nfs_rpc_free(p0);
 	return -nfs_stat_to_errno(status);
 }
 
@@ -281,17 +339,20 @@ int nfs_proc_read(struct nfs_server *server, struct nfs_fh *fhandle,
 {
 	int *p, *p0;
 	int status;
+	int ruid = 0;
 	int len = 0; /* = 0 is for gcc */
 
 	PRINTK("NFS call  read %d @ %d\n", count, offset);
-	p = p0 = (int *) get_free_page(GFP_KERNEL);
-	p = nfs_rpc_header(p, NFSPROC_READ);
+	if (!(p0 = nfs_rpc_alloc()))
+		return -EIO;
+retry:
+	p = nfs_rpc_header(p0, NFSPROC_READ, ruid);
 	p = xdr_encode_fhandle(p, fhandle);
 	*p++ = htonl(offset);
 	*p++ = htonl(count);
 	*p++ = htonl(count); /* traditional, could be any value */
 	if ((status = nfs_rpc_call(server, p0, p)) < 0) {
-		free_page((long) p0);
+		nfs_rpc_free(p0);
 		return status;
 	}
 	if (!(p = nfs_rpc_verify(p0)))
@@ -305,9 +366,14 @@ int nfs_proc_read(struct nfs_server *server, struct nfs_fh *fhandle,
 		else
 			PRINTK("NFS reply read %d\n", len);
 	}
-	else
+	else {
+		if (!ruid && current->euid == 0 && current->uid != 0) {
+			ruid = 1;
+			goto retry;
+		}
 		PRINTK("NFS reply read failed = %d\n", status);
-	free_page((long) p0);
+	}
+	nfs_rpc_free(p0);
 	return (status == NFS_OK) ? len : -nfs_stat_to_errno(status);
 }
 
@@ -316,17 +382,20 @@ int nfs_proc_write(struct nfs_server *server, struct nfs_fh *fhandle,
 {
 	int *p, *p0;
 	int status;
+	int ruid = 0;
 
 	PRINTK("NFS call  write %d @ %d\n", count, offset);
-	p = p0 = (int *) get_free_page(GFP_KERNEL);
-	p = nfs_rpc_header(p, NFSPROC_WRITE);
+	if (!(p0 = nfs_rpc_alloc()))
+		return -EIO;
+retry:
+	p = nfs_rpc_header(p0, NFSPROC_WRITE, ruid);
 	p = xdr_encode_fhandle(p, fhandle);
 	*p++ = htonl(offset); /* traditional, could be any value */
 	*p++ = htonl(offset);
 	*p++ = htonl(count); /* traditional, could be any value */
 	p = xdr_encode_data(p, data, count);
 	if ((status = nfs_rpc_call(server, p0, p)) < 0) {
-		free_page((long) p0);
+		nfs_rpc_free(p0);
 		return status;
 	}
 	if (!(p = nfs_rpc_verify(p0)))
@@ -335,9 +404,14 @@ int nfs_proc_write(struct nfs_server *server, struct nfs_fh *fhandle,
 		p = xdr_decode_fattr(p, fattr);
 		PRINTK("NFS reply write\n");
 	}
-	else
+	else {
+		if (!ruid && current->euid == 0 && current->uid != 0) {
+			ruid = 1;
+			goto retry;
+		}
 		PRINTK("NFS reply write failed = %d\n", status);
-	free_page((long) p0);
+	}
+	nfs_rpc_free(p0);
 	return -nfs_stat_to_errno(status);
 }
 
@@ -347,15 +421,18 @@ int nfs_proc_create(struct nfs_server *server, struct nfs_fh *dir,
 {
 	int *p, *p0;
 	int status;
+	int ruid = 0;
 
 	PRINTK("NFS call  create %s\n", name);
-	p = p0 = (int *) get_free_page(GFP_KERNEL);
-	p = nfs_rpc_header(p, NFSPROC_CREATE);
+	if (!(p0 = nfs_rpc_alloc()))
+		return -EIO;
+retry:
+	p = nfs_rpc_header(p0, NFSPROC_CREATE, ruid);
 	p = xdr_encode_fhandle(p, dir);
 	p = xdr_encode_string(p, name);
 	p = xdr_encode_sattr(p, sattr);
 	if ((status = nfs_rpc_call(server, p0, p)) < 0) {
-		free_page((long) p0);
+		nfs_rpc_free(p0);
 		return status;
 	}
 	if (!(p = nfs_rpc_verify(p0)))
@@ -365,9 +442,14 @@ int nfs_proc_create(struct nfs_server *server, struct nfs_fh *dir,
 		p = xdr_decode_fattr(p, fattr);
 		PRINTK("NFS reply create\n");
 	}
-	else
+	else {
+		if (!ruid && current->euid == 0 && current->uid != 0) {
+			ruid = 1;
+			goto retry;
+		}
 		PRINTK("NFS reply create failed = %d\n", status);
-	free_page((long) p0);
+	}
+	nfs_rpc_free(p0);
 	return -nfs_stat_to_errno(status);
 }
 
@@ -375,14 +457,17 @@ int nfs_proc_remove(struct nfs_server *server, struct nfs_fh *dir, const char *n
 {
 	int *p, *p0;
 	int status;
+	int ruid = 0;
 
 	PRINTK("NFS call  remove %s\n", name);
-	p = p0 = (int *) get_free_page(GFP_KERNEL);
-	p = nfs_rpc_header(p, NFSPROC_REMOVE);
+	if (!(p0 = nfs_rpc_alloc()))
+		return -EIO;
+retry:
+	p = nfs_rpc_header(p0, NFSPROC_REMOVE, ruid);
 	p = xdr_encode_fhandle(p, dir);
 	p = xdr_encode_string(p, name);
 	if ((status = nfs_rpc_call(server, p0, p)) < 0) {
-		free_page((long) p0);
+		nfs_rpc_free(p0);
 		return status;
 	}
 	if (!(p = nfs_rpc_verify(p0)))
@@ -390,9 +475,14 @@ int nfs_proc_remove(struct nfs_server *server, struct nfs_fh *dir, const char *n
 	else if ((status = ntohl(*p++)) == NFS_OK) {
 		PRINTK("NFS reply remove\n");
 	}
-	else
+	else {
+		if (!ruid && current->euid == 0 && current->uid != 0) {
+			ruid = 1;
+			goto retry;
+		}
 		PRINTK("NFS reply remove failed = %d\n", status);
-	free_page((long) p0);
+	}
+	nfs_rpc_free(p0);
 	return -nfs_stat_to_errno(status);
 }
 
@@ -402,16 +492,19 @@ int nfs_proc_rename(struct nfs_server *server,
 {
 	int *p, *p0;
 	int status;
+	int ruid = 0;
 
 	PRINTK("NFS call  rename %s -> %s\n", old_name, new_name);
-	p = p0 = (int *) get_free_page(GFP_KERNEL);
-	p = nfs_rpc_header(p, NFSPROC_RENAME);
+	if (!(p0 = nfs_rpc_alloc()))
+		return -EIO;
+retry:
+	p = nfs_rpc_header(p0, NFSPROC_RENAME, ruid);
 	p = xdr_encode_fhandle(p, old_dir);
 	p = xdr_encode_string(p, old_name);
 	p = xdr_encode_fhandle(p, new_dir);
 	p = xdr_encode_string(p, new_name);
 	if ((status = nfs_rpc_call(server, p0, p)) < 0) {
-		free_page((long) p0);
+		nfs_rpc_free(p0);
 		return status;
 	}
 	if (!(p = nfs_rpc_verify(p0)))
@@ -419,9 +512,14 @@ int nfs_proc_rename(struct nfs_server *server,
 	else if ((status = ntohl(*p++)) == NFS_OK) {
 		PRINTK("NFS reply rename\n");
 	}
-	else
+	else {
+		if (!ruid && current->euid == 0 && current->uid != 0) {
+			ruid = 1;
+			goto retry;
+		}
 		PRINTK("NFS reply rename failed = %d\n", status);
-	free_page((long) p0);
+	}
+	nfs_rpc_free(p0);
 	return -nfs_stat_to_errno(status);
 }
 
@@ -430,15 +528,18 @@ int nfs_proc_link(struct nfs_server *server, struct nfs_fh *fhandle,
 {
 	int *p, *p0;
 	int status;
+	int ruid = 0;
 
 	PRINTK("NFS call  link %s\n", name);
-	p = p0 = (int *) get_free_page(GFP_KERNEL);
-	p = nfs_rpc_header(p, NFSPROC_LINK);
+	if (!(p0 = nfs_rpc_alloc()))
+		return -EIO;
+retry:
+	p = nfs_rpc_header(p0, NFSPROC_LINK, ruid);
 	p = xdr_encode_fhandle(p, fhandle);
 	p = xdr_encode_fhandle(p, dir);
 	p = xdr_encode_string(p, name);
 	if ((status = nfs_rpc_call(server, p0, p)) < 0) {
-		free_page((long) p0);
+		nfs_rpc_free(p0);
 		return status;
 	}
 	if (!(p = nfs_rpc_verify(p0)))
@@ -446,9 +547,14 @@ int nfs_proc_link(struct nfs_server *server, struct nfs_fh *fhandle,
 	else if ((status = ntohl(*p++)) == NFS_OK) {
 		PRINTK("NFS reply link\n");
 	}
-	else
+	else {
+		if (!ruid && current->euid == 0 && current->uid != 0) {
+			ruid = 1;
+			goto retry;
+		}
 		PRINTK("NFS reply link failed = %d\n", status);
-	free_page((long) p0);
+	}
+	nfs_rpc_free(p0);
 	return -nfs_stat_to_errno(status);
 }
 
@@ -457,16 +563,19 @@ int nfs_proc_symlink(struct nfs_server *server, struct nfs_fh *dir,
 {
 	int *p, *p0;
 	int status;
+	int ruid = 0;
 
 	PRINTK("NFS call  symlink %s -> %s\n", name, path);
-	p = p0 = (int *) get_free_page(GFP_KERNEL);
-	p = nfs_rpc_header(p, NFSPROC_SYMLINK);
+	if (!(p0 = nfs_rpc_alloc()))
+		return -EIO;
+retry:
+	p = nfs_rpc_header(p0, NFSPROC_SYMLINK, ruid);
 	p = xdr_encode_fhandle(p, dir);
 	p = xdr_encode_string(p, name);
 	p = xdr_encode_string(p, path);
 	p = xdr_encode_sattr(p, sattr);
 	if ((status = nfs_rpc_call(server, p0, p)) < 0) {
-		free_page((long) p0);
+		nfs_rpc_free(p0);
 		return status;
 	}
 	if (!(p = nfs_rpc_verify(p0)))
@@ -474,9 +583,14 @@ int nfs_proc_symlink(struct nfs_server *server, struct nfs_fh *dir,
 	else if ((status = ntohl(*p++)) == NFS_OK) {
 		PRINTK("NFS reply symlink\n");
 	}
-	else
+	else {
+		if (!ruid && current->euid == 0 && current->uid != 0) {
+			ruid = 1;
+			goto retry;
+		}
 		PRINTK("NFS reply symlink failed = %d\n", status);
-	free_page((long) p0);
+	}
+	nfs_rpc_free(p0);
 	return -nfs_stat_to_errno(status);
 }
 
@@ -486,15 +600,18 @@ int nfs_proc_mkdir(struct nfs_server *server, struct nfs_fh *dir,
 {
 	int *p, *p0;
 	int status;
+	int ruid = 0;
 
 	PRINTK("NFS call  mkdir %s\n", name);
-	p = p0 = (int *) get_free_page(GFP_KERNEL);
-	p = nfs_rpc_header(p, NFSPROC_MKDIR);
+	if (!(p0 = nfs_rpc_alloc()))
+		return -EIO;
+retry:
+	p = nfs_rpc_header(p0, NFSPROC_MKDIR, ruid);
 	p = xdr_encode_fhandle(p, dir);
 	p = xdr_encode_string(p, name);
 	p = xdr_encode_sattr(p, sattr);
 	if ((status = nfs_rpc_call(server, p0, p)) < 0) {
-		free_page((long) p0);
+		nfs_rpc_free(p0);
 		return status;
 	}
 	if (!(p = nfs_rpc_verify(p0)))
@@ -504,9 +621,14 @@ int nfs_proc_mkdir(struct nfs_server *server, struct nfs_fh *dir,
 		p = xdr_decode_fattr(p, fattr);
 		PRINTK("NFS reply mkdir\n");
 	}
-	else
+	else {
+		if (!ruid && current->euid == 0 && current->uid != 0) {
+			ruid = 1;
+			goto retry;
+		}
 		PRINTK("NFS reply mkdir failed = %d\n", status);
-	free_page((long) p0);
+	}
+	nfs_rpc_free(p0);
 	return -nfs_stat_to_errno(status);
 }
 
@@ -514,14 +636,17 @@ int nfs_proc_rmdir(struct nfs_server *server, struct nfs_fh *dir, const char *na
 {
 	int *p, *p0;
 	int status;
+	int ruid = 0;
 
 	PRINTK("NFS call  rmdir %s\n", name);
-	p = p0 = (int *) get_free_page(GFP_KERNEL);
-	p = nfs_rpc_header(p, NFSPROC_RMDIR);
+	if (!(p0 = nfs_rpc_alloc()))
+		return -EIO;
+retry:
+	p = nfs_rpc_header(p0, NFSPROC_RMDIR, ruid);
 	p = xdr_encode_fhandle(p, dir);
 	p = xdr_encode_string(p, name);
 	if ((status = nfs_rpc_call(server, p0, p)) < 0) {
-		free_page((long) p0);
+		nfs_rpc_free(p0);
 		return status;
 	}
 	if (!(p = nfs_rpc_verify(p0)))
@@ -529,9 +654,14 @@ int nfs_proc_rmdir(struct nfs_server *server, struct nfs_fh *dir, const char *na
 	else if ((status = ntohl(*p++)) == NFS_OK) {
 		PRINTK("NFS reply rmdir\n");
 	}
-	else
+	else {
+		if (!ruid && current->euid == 0 && current->uid != 0) {
+			ruid = 1;
+			goto retry;
+		}
 		PRINTK("NFS reply rmdir failed = %d\n", status);
-	free_page((long) p0);
+	}
+	nfs_rpc_free(p0);
 	return -nfs_stat_to_errno(status);
 }
 
@@ -540,19 +670,22 @@ int nfs_proc_readdir(struct nfs_server *server, struct nfs_fh *fhandle,
 {
 	int *p, *p0;
 	int status;
+	int ruid = 0;
 	int i = 0; /* = 0 is for gcc */
 	int size;
 	int eof;
 
 	PRINTK("NFS call  readdir %d @ %d\n", count, cookie);
 	size = server->rsize;
-	p = p0 = (int *) get_free_page(GFP_KERNEL);
-	p = nfs_rpc_header(p, NFSPROC_READDIR);
+	if (!(p0 = nfs_rpc_alloc()))
+		return -EIO;
+retry:
+	p = nfs_rpc_header(p0, NFSPROC_READDIR, ruid);
 	p = xdr_encode_fhandle(p, fhandle);
 	*p++ = htonl(cookie);
 	*p++ = htonl(size);
 	if ((status = nfs_rpc_call(server, p0, p)) < 0) {
-		free_page((long) p0);
+		nfs_rpc_free(p0);
 		return status;
 	}
 	if (!(p = nfs_rpc_verify(p0)))
@@ -575,9 +708,14 @@ int nfs_proc_readdir(struct nfs_server *server, struct nfs_fh *fhandle,
 			       eof ? "eof" : "");
 		}
 	}
-	else
+	else {
+		if (!ruid && current->euid == 0 && current->uid != 0) {
+			ruid = 1;
+			goto retry;
+		}
 		PRINTK("NFS reply readdir failed = %d\n", status);
-	free_page((long) p0);
+	}
+	nfs_rpc_free(p0);
 	return (status == NFS_OK) ? i : -nfs_stat_to_errno(status);
 }
 
@@ -586,13 +724,16 @@ int nfs_proc_statfs(struct nfs_server *server, struct nfs_fh *fhandle,
 {
 	int *p, *p0;
 	int status;
+	int ruid = 0;
 
 	PRINTK("NFS call  statfs\n");
-	p = p0 = (int *) get_free_page(GFP_KERNEL);
-	p = nfs_rpc_header(p, NFSPROC_STATFS);
+	if (!(p0 = nfs_rpc_alloc()))
+		return -EIO;
+retry:
+	p = nfs_rpc_header(p0, NFSPROC_STATFS, ruid);
 	p = xdr_encode_fhandle(p, fhandle);
 	if ((status = nfs_rpc_call(server, p0, p)) < 0) {
-		free_page((long) p0);
+		nfs_rpc_free(p0);
 		return status;
 	}
 	if (!(p = nfs_rpc_verify(p0)))
@@ -601,9 +742,14 @@ int nfs_proc_statfs(struct nfs_server *server, struct nfs_fh *fhandle,
 		p = xdr_decode_fsinfo(p, res);
 		PRINTK("NFS reply statfs\n");
 	}
-	else
+	else {
+		if (!ruid && current->euid == 0 && current->uid != 0) {
+			ruid = 1;
+			goto retry;
+		}
 		PRINTK("NFS reply statfs failed = %d\n", status);
-	free_page((long) p0);
+	}
+	nfs_rpc_free(p0);
 	return -nfs_stat_to_errno(status);
 }
 
@@ -611,7 +757,7 @@ int nfs_proc_statfs(struct nfs_server *server, struct nfs_fh *fhandle,
  * Here are a few RPC-assist functions.
  */
 
-static int *nfs_rpc_header(int *p, int procedure)
+static int *nfs_rpc_header(int *p, int procedure, int ruid)
 {
 	int *p1, *p2;
 	int i;
@@ -632,7 +778,7 @@ static int *nfs_rpc_header(int *p, int procedure)
 	p1 = p++;
 	*p++ = htonl(CURRENT_TIME); /* traditional, could be anything */
 	p = xdr_encode_string(p, (char *) sys);
-	*p++ = htonl(current->euid);
+	*p++ = htonl(ruid ? current->uid : current->euid);
 	*p++ = htonl(current->egid);
 	p2 = p++;
 	for (i = 0; i < 16 && i < NGROUPS && current->groups[i] != NOGROUP; i++)
@@ -686,7 +832,7 @@ static int *nfs_rpc_verify(int *p)
 #endif
 
 static struct {
-	enum nfs_stat stat;
+	int stat;
 	int errno;
 } nfs_errtbl[] = {
 	{ NFS_OK,		0		},
@@ -699,6 +845,7 @@ static struct {
 	{ NFSERR_NODEV,		ENODEV		},
 	{ NFSERR_NOTDIR,	ENOTDIR		},
 	{ NFSERR_ISDIR,		EISDIR		},
+	{ NFSERR_INVAL,		EINVAL		},
 	{ NFSERR_FBIG,		EFBIG		},
 	{ NFSERR_NOSPC,		ENOSPC		},
 	{ NFSERR_ROFS,		EROFS		},
